Revert "change links"
