+++ index.html
@@ -51,17 +51,17 @@
       <nav class="navbar navbar-expand-lg navbar-light bg-light">
         <ul class="navbar-nav">
           <li class="nav-item">
-            <a id="main-page" class="nav-link" href="/web-bee-internship-01/"
+            <a id="main-page" class="nav-link" href="/"
               ><i class="fas fa-user mr-2"></i>Profile</a
             >
           </li>
           <li class="nav-item">
-            <a class="nav-link" href="/web-bee-internship-01/map"
+            <a class="nav-link" href="/map"
               ><i class="fas fa-map-marked-alt mr-2"></i>Map</a
             >
           </li>
           <li class="nav-item">
-            <a class="nav-link" href="/web-bee-internship-01/time"
+            <a class="nav-link" href="/time"
               ><i class="fas fa-clock mr-2"></i>Time</a
             >
           </li>
