+++ package.json
@@ -14,5 +14,8 @@
     "gh-pages": "6.1.1",
     "typescript": "^5.2.2",
     "vite": "^5.2.0"
+  },
+  "vite": {
+    "base": "/web-bee-internship-01"
   }
 }
