+++ src/router/index.ts
@@ -1,30 +1,21 @@
-import {
-  setupPage,
-  updateNavBarClassList,
-  updateStoredActiveLink,
-  restoreActiveLink,
-} from "../utils";
+import { setupPage } from "../utils";
 import { renderMap } from "../map";
+import { updateNavBar, restoreActiveLink } from "../utils/nav-bar";
 
 const routes: Routes = {
   paths: {
-    "/web-bee-internship-01/404": "/public/404.html",
-    "/web-bee-internship-01/": "/public/profile.html",
-    "/web-bee-internship-01/map": "/public/map.html",
-    "/web-bee-internship-01/time": "/public/timer.html",
+    "404": "/public/404.html",
+    "/": "/public/profile.html",
+    "/map": "/public/map.html",
+    "/time": "/public/timer.html",
   },
   titles: {
-    "/web-bee-internship-01/404": "Page not found",
-    "/web-bee-internship-01/": "Profile",
-    "/web-bee-internship-01/map": "Map",
-    "/web-bee-internship-01/time": "Time",
+    "404": "Page not found",
+    "/": "Profile",
+    "/map": "Map",
+    "/time": "Time",
   },
 };
-
-document.addEventListener("DOMContentLoaded", function () {
-  window.history.replaceState({}, "", "/web-bee-internship-01/");
-  handleLocation();
-});
 
 const handleLocation = async () => {
   const path = window.location.pathname;
@@ -40,8 +31,7 @@ const navigateTo = (event: Event) => {
   event.preventDefault();
   const target = event.target as HTMLAnchorElement;
   window.history.pushState({}, "", target.href);
-  updateNavBarClassList(target);
-  updateStoredActiveLink(target);
+  updateNavBar(target);
   handleLocation();
 };
