+++ src/timer/index.ts
@@ -41,3 +41,6 @@ const padZero = (value: number) => value.toString().padStart(2, "0");
 const updateStoredTimer = (elapsedTime: number) => {
   sessionStorage.setItem("timerValue", elapsedTime.toString());
 };
+
+export const renderTimer = (path: string) =>
+  path === "/web-bee-internship-01/time" && startTimer();
