+++ src/utils/nav-bar.ts
@@ -1,0 +1,31 @@
+function updateNavBarClassList(target: HTMLAnchorElement) {
+  document.querySelectorAll(".nav-link").forEach((link) => {
+    link.classList.remove("active");
+  });
+  target.classList.add("active");
+}
+
+function updateStoredActiveLink(target: HTMLAnchorElement) {
+  sessionStorage.setItem("activeLink", target.href);
+}
+
+function restoreActiveLink() {
+  const activeLink = sessionStorage.getItem("activeLink");
+  if (activeLink) {
+    document.querySelectorAll(".nav-link").forEach((element) => {
+      const link = element as HTMLAnchorElement;
+      if (link.href === activeLink) {
+        link.classList.add("active");
+      }
+    });
+  } else {
+    document.querySelector("#main-page")?.classList.add("active");
+  }
+}
+
+function updateNavBar(target: HTMLAnchorElement) {
+  updateNavBarClassList(target);
+  updateStoredActiveLink(target);
+}
+
+export { updateNavBar, restoreActiveLink };
